modularized perms class
